Changes for snap-1450
