+++ cluster/src/main/scala/io/snappydata/gemxd/ClusterCallbacksImpl.scala
@@ -24,7 +24,7 @@ import com.gemstone.gemfire.internal.shared.Version
 import com.pivotal.gemfirexd.internal.iapi.types.DataValueDescriptor
 import com.pivotal.gemfirexd.internal.impl.sql.execute.ValueRow
 import com.pivotal.gemfirexd.internal.snappy.{CallbackFactoryProvider, ClusterCallbacks, LeadNodeExecutionContext, SparkSQLExecute}
-import io.snappydata.SnappyTableStatsProviderService
+import io.snappydata.{SnappyEmbeddedTableStatsProviderService, SnappyTableStatsProviderService}
 import io.snappydata.cluster.ExecutorInitiator
 import io.snappydata.impl.LeadImpl
 
@@ -108,6 +108,6 @@ object ClusterCallbacksImpl extends ClusterCallbacks with Logging {
   }
 
   override def publishColumnTableStats(): Unit = {
-    SnappyTableStatsProviderService.publishColumnTableRowCountStats()
+    SnappyEmbeddedTableStatsProviderService.publishColumnTableRowCountStats()
   }
 }
