+++ cluster/src/main/scala/org/apache/spark/ui/SnappyStatsPage.scala
@@ -36,7 +36,7 @@ private[ui] class SnappyStatsPage(parent: SnappyStatsTab)
   val numFormatter = java.text.NumberFormat.getIntegerInstance
 
   def render(request: HttpServletRequest): Seq[Node] = {
-    val uiDisplayInfo = SnappyTableStatsProviderService
+    val uiDisplayInfo = SnappyTableStatsProviderService.getService
         .getAggregatedStatsOnDemand
 
     val uiTableInfo = uiDisplayInfo._1
