+++ core/src/main/scala/org/apache/spark/sql/SmartConnectorHelper.scala
@@ -75,7 +75,7 @@ class SmartConnectorHelper(snappySession: SnappySession) extends Logging {
     try {
       function
     } catch {
-      case e: SQLException if isConnectionException(e) =>
+      case e: SQLException if SmartConnectorHelper.isConnectionException(e) =>
         // attempt to create a new connection if connection
         // is closed
         SmartConnectorHelper.close()
@@ -84,10 +84,7 @@ class SmartConnectorHelper(snappySession: SnappySession) extends Logging {
     }
   }
 
-  private def isConnectionException(e: SQLException): Boolean = {
-    e.getSQLState.startsWith(SQLState.CONNECTIVITY_PREFIX) ||
-    e.getSQLState.startsWith(SQLState.LANG_DEAD_STATEMENT)
-  }
+
 
   def createTable(
       tableIdent: QualifiedTableName,
@@ -297,5 +294,10 @@ object SmartConnectorHelper {
     val baip = new ByteArrayInputStream(value)
     val ois = new ObjectInputStream(baip)
     ois.readObject()
+  }
+
+  def isConnectionException(e: SQLException): Boolean = {
+    e.getSQLState.startsWith(SQLState.CONNECTIVITY_PREFIX) ||
+        e.getSQLState.startsWith(SQLState.LANG_DEAD_STATEMENT)
   }
 }
