+++ core/src/main/scala/org/apache/spark/sql/execution/columnar/JDBCAppendableRelation.scala
@@ -19,7 +19,7 @@ package org.apache.spark.sql.execution.columnar
 import java.sql.Connection
 import java.util.concurrent.locks.ReentrantReadWriteLock
 
-import _root_.io.snappydata.{Constant, SnappyTableStatsProviderService}
+import io.snappydata.{SnappyThinConnectorTableStatsProvider, Constant, SnappyTableStatsProviderService}
 
 import org.apache.spark.Logging
 import org.apache.spark.rdd.RDD
@@ -74,12 +74,14 @@ abstract case class JDBCAppendableRelation(
   def numBuckets: Int = -1
 
   override def sizeInBytes: Long = {
-    SnappyTableStatsProviderService.getTableStatsFromService(table) match {
+    logInfo(s"JDBCAppendableRelation.sizeInBytes sdeshmukh looking size of table $table")
+    val size = SnappyTableStatsProviderService.getService.getTableStatsFromService(table) match {
       case Some(s) => s.getTotalSize
       case None => super.sizeInBytes
     }
+    logInfo(s"JDBCAppendableRelation.sizeInBytes sdeshmukh $table size is $size")
+    size
   }
-
 
   protected final def dialect: JdbcDialect = connProperties.dialect
